Remove cyclic dependency between Geometry/HGCalGeometry and Geometry/CaloTopology
