+++ Geometry/HGCalGeometry/test/BuildFile.xml
@@ -6,6 +6,7 @@
 <use name="FWCore/ServiceRegistry"/>
 <use name="Geometry/HcalTowerAlgo"/>
 <use name="Geometry/HGCalGeometry"/>
+<use name="Geometry/CaloTopology"/>
 <use name="Geometry/Records"/>
 <use name="MagneticField/Engine"/>
 <use name="MagneticField/Records"/>
