+++ Geometry/HGCalGeometry/test/testHGCalNeighbour_cfg.py
@@ -55,7 +55,7 @@ process.load(geomFile)
 process.load("SimGeneral.HepPDTESSource.pdt_cfi")
 process.load('Configuration.StandardSequences.MagneticField_cff')
 process.load('Configuration.StandardSequences.Services_cff')
-process.load('Geometry.CaloTopology.hgcalNeighbourTester_cfi')
+process.load('Geometry.HGCalGeometry.hgcalNeighbourTester_cfi')
 process.load('FWCore.MessageService.MessageLogger_cfi')
 process.load('Configuration.EventContent.EventContent_cff')
 process.load('Configuration.StandardSequences.FrontierConditions_GlobalTag_cff')
